# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "学员端(hrwai_user) 学习/练习/考试/收藏/搜索/资料/论坛等 P0/P1 能力（ADR-0017/0018）。鉴权：`Authorization: Bearer \u003caccess JWT\u003e`（access 2h，`POST /api/auth/refresh` 换新双令牌）。响应统一 `{code,message,data}`（见 ADR-0005）。本页由 gin-swagger 生成，范围 B（~30 端点）。",
        "title": "叉车维修培训系统-学员端 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/practice-mode/free": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按题型随机抽题，count 控制题量（默认 20）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-练习"
                ],
                "summary": "随机练习抽题",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "题量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "题型 single_choice 等",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/forklift-training_internal_service.QuestionDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/practice-mode/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页查询练习历史，支持按题型/日期过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-练习"
                ],
                "summary": "练习历史",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "题型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期 YYYY-MM-DD",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 YYYY-MM-DD",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.HistoryResultDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/practice-mode/progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按 mode 查询断点续练进度；mode 为空默认为 sequential",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-练习"
                ],
                "summary": "查询练习进度",
                "parameters": [
                    {
                        "type": "string",
                        "default": "sequential",
                        "description": "练习模式",
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.ProgressResultDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "支持顺序/标签等练习的断点续练",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-练习"
                ],
                "summary": "保存练习游标与答题状态",
                "parameters": [
                    {
                        "description": "进度",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/practice-mode/sequential": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回当前批次题目 + 进度",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-练习"
                ],
                "summary": "顺序练习开始/续练",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.PracticeStartResultDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/practice-mode/sequential-progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用于卡片展示的进度快照",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-练习"
                ],
                "summary": "顺序练习进度",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.ProgressResultDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/practice-mode/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "汇总练习正确率/已练题量等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-练习"
                ],
                "summary": "练习统计",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.PracticeStatsDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/practice-mode/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "提交单题答案并即时判分",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-练习"
                ],
                "summary": "提交练习答案",
                "parameters": [
                    {
                        "description": "答题",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.SubmitResultDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/practice-mode/tag": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按标签 ID 开始/续练专项练习；count=0 表示全部",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-练习"
                ],
                "summary": "标签专项练习",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "题库标签ID",
                        "name": "tag_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "题量 0=全部",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.PracticeStartResultDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "forklift-training_internal_service.HistoryResultDTO": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forklift-training_internal_service.HistoryItemDTO"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "forklift-training_internal_service.QuestionDTO": {
            "type": "object",
            "properties": {
                "answer": {
                    "description": "学员侧（includeAnswer=false）省略以下四个字段。",
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "created_by_type": {
                    "type": "string"
                },
                "explanation": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "options": {},
                "reference_answer": {
                    "type": "string"
                },
                "reject_reason": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "scoring_criteria": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "description": "Tags 题库管理面附加（未设置时省略；设置后保留 null/[] 形态与历史一致）。"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "forklift-training_pkg_response.R": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003caccess JWT\u003e，示例：Bearer eyJhbGciOi...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}